% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Create_graphics.R
\name{create_final_analysis_graph}
\alias{create_final_analysis_graph}
\title{Final analizimiz için bar grafik oluşturma}
\usage{
create_final_analysis_graph(
  data,
  x,
  y,
  fill,
  title,
  x_label,
  y_label,
  fill_label,
  angle = 30
)
}
\arguments{
\item{data}{Veri çerçevesi.}

\item{x}{X ekseni için kullanılacak değişken adı.}

\item{y}{Y ekseni için kullanılacak değişken adı.}

\item{fill}{Barların dolgu estetiği için kullanılacak değişken adı.}

\item{title}{Grafiğin başlığı.}

\item{x_label}{X ekseninin etiketi.}

\item{y_label}{Y ekseninin etiketi.}

\item{fill_label}{Dolgu estetiğinin etiketi.}

\item{angle}{X ekseni metinlerinin açısı (varsayılan 30 derece).}
}
\value{
ggplot2 nesnesi olarak döndürür.
}
\description{
Verilen veri çerçevesine dayalı olarak bir bar plot oluşturur.
}
