% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Final_analiz.R
\name{filter_matched_categories}
\alias{filter_matched_categories}
\title{Filter matched categories based on total spending}
\usage{
filter_matched_categories(data1, data2, group_cols)
}
\arguments{
\item{group_cols}{Eslestirme ypinca hangi columnlar alınsın ve geriye dondurulsun}

\item{result}{Eslestirme yapilacak dataframe1}

\item{maxResult}{Eslestirme yapilacak dataframe1}
}
\value{
Eslesen veri cercevelerini dondurur
}
\description{
This function filters the matched categories based on the total spending
being in the maximum total spending.
}
