% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Csv_operations.R
\name{process_data}
\alias{process_data}
\title{Veriyi yükler, eksik verileri işler, gereksiz kolonları kaldırır ve sonuçları kaydeder}
\usage{
process_data(data, output_path, sep = ";", remove_cols)
}
\arguments{
\item{output_path}{Çıkış CSV dosyasının yolu.}

\item{sep}{CSV dosyasındaki ayırıcı karakter (varsayılan ";").}

\item{remove_cols}{Kaldırılacak kolonların isimleri.}

\item{input_path}{Giriş CSV dosyasının yolu.}
}
\description{
Veriyi yükler, eksik verileri işler, gereksiz kolonları kaldırır ve sonuçları kaydeder
}
