% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Csv_operations.R
\name{remove_unnecessary_columns}
\alias{remove_unnecessary_columns}
\title{Gereksiz kolonları kaldırır}
\usage{
remove_unnecessary_columns(data, columns)
}
\arguments{
\item{data}{Veri çerçevesi.}

\item{columns}{Kaldırılacak kolonların isimleri.}
}
\value{
Düzenlenmiş veri çerçevesi.
}
\description{
Gereksiz kolonları kaldırır
}
