% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Csv_operations.R
\name{summarize_data}
\alias{summarize_data}
\title{Veriye genel bakış ve eksik verileri kontrol eder}
\usage{
summarize_data(data)
}
\arguments{
\item{data}{Veri çerçevesi.}
}
\description{
Veriye genel bakış ve eksik verileri kontrol eder
}
